remove incorrectly named file
